backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
apispec==6.6.1
attrs==24.2.0
base58==2.0.1
bech32==1.2.0
blinker==1.8.2
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
cosmpy==0.9.2
ecdsa==0.15
fetchai-ledger-api
Flask==3.0.3
frozenlist==1.4.1
googleapis-common-protos==1.65.0
grpcio==1.66.2
h11==0.14.0
idna==2.8
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lark-parser==0.8.1
MarkupSafe==3.0.1
msgpack==0.6.2
multidict==6.1.0
packaging==24.1
propcache==0.2.0
protobuf==4.25.5
pyaes==1.6.1
pycryptodome==3.21.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
referencing==0.35.1
requests==2.22.0
rpds-py==0.20.0
semver==2.10.2
six==1.16.0
typing_extensions==4.12.2
uagents
urllib3==1.25.11
uvicorn==0.30.6
websockets==10.4
Werkzeug==3.0.4
yarl==1.15.2
